Create "swing-lib" subproject and move `swinglib` package there
