+++ settings.gradle
@@ -6,4 +6,5 @@ include 'http-client'
 include 'http-server'
 include 'lobby'
 include 'lobby-db'
+include 'swing-lib'
 include 'test-common'
