+++ swing-lib/build.gradle
@@ -1,0 +1,5 @@
+description = 'TripleA library for utilities and components that depend only on core java, javax.swing, and java.awt'
+
+dependencies {
+    testImplementation project(':test-common')
+}
